requirements.txt
accelerate==0.33.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==23.2.0
beam-client==0.2.96
beta9==0.1.97
betterproto-beta9==2.0.0b7
black==24.10.0
boto3==1.35.44
botocore==1.35.44
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
cog==0.11.6
croniter==3.0.4
diffusers @ git+https://github.com/huggingface/diffusers.git@b821f006d025b72654223b32e6bf0c49afed371d
fastapi==0.110.3
filelock==3.16.1
fsspec==2024.10.0
grpcio==1.67.1
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httptools==0.6.4
huggingface-hub==0.26.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
isort==5.13.2
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
ninja==1.11.1.1
numpy==1.26.4
optimum-quanto==0.2.5
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
protobuf==4.25.5
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
s3transfer==0.10.3
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
structlog==24.4.0
sympy==1.13.1
tokenizers==0.19.1
torch==2.5.0
tqdm==4.66.5
transformers==4.43.3
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.21.0
watchdog==4.0.2
watchfiles==0.24.0
websockets==12.0
zipp==3.20.2
